So the hue-bucket analysis was no good. Reverting to working version.
The issue here was that PaletteReducer.analyze() has an overload for one Pixmap, and another for a sequence of Pixmaps. When AnimatedGif does this, it always calls the second, but PNG8 can call either. Only the first was edited to use the hue-bucket technique, so only still PNG images showed the (terrible) changes.

This commit also makes PaletteReducer.reduceBlueNoise() act like the other blue noise reductions in PNG8 and AnimatedGif