More lightness changes; we don't need the cludge on ditherStrength anymore.
If you setDitherStrength(1f), the result should be pretty close to what it was before in terms of lightness. It should look better than it did before for values less than 1.0, with a more "tactical" use of less dithering, and will, as one would expect, act weird when given values much higher than 1. It does look fine for some values between 1 and about 1.5, and this could be desirable.