Get QualityPalette in the right range.
It was producing differences that were generally between 10x and 20x where they should have been, and this affected the analyze() code. It meant that many less-common colors would be omitted from the palette, and the bulk of colors could be similar. Now PaletteReducer should be similar to FastPalette and QualityPalette.