Test all the analysis methods...
The "Ocean" shader loop seems to analyze incredibly badly by default... Reductive does OK with it, but HueWise and the regular analyze() seem to draw a false similarity between blue and red, and place the two next to each other when it doesn't make sense. They might just be omitting violet and purple, and trying to fill in with red and blue.