Default thresholds really needed work.
The thresholds were way too low for analyze() and analyzeHueWise(), and a little low for analyzeReductive (but one place set them far too low as well...). This affected the quality of animations, mainly, if they didn't use the default palette (or any exact palette). The current thresholds should behave better, though we'll have to see.