BLUNT dither isn't the best, but it's an option!
BLUE_NOISE can look better than BLUNT on many images, and small palettes likely behave quite differently from large ones. BLUE_NOISE always applies a strong checkerboard, even for large palettes, so that may be a noticeable artifact in that case. BLUNT also applies a checkerboard, but it's normally about a quarter the strength as the one in BLUE_NOISE.